Moved resetting lastFeatherCount and reachedSummit to start block.
The timer doesn't reset after a run is finished, it must be done manually. This means that lastFeatherCount and reachedSummit stay the same in the very first run after the finished run, since the reset block is never activated. Moving this to the start block should fix the issue.
Also removed the overarching "Choose your splits here" setting, since it was unnecessary.